+++ src/main/java/firok/tiths/common/Blocks.java
@@ -86,6 +86,9 @@ public class Blocks
 	@Reg(Keys.oreTopaz)
 	@GenOreWorld(times=2,timeRate = 0.8f,size=4,minY = 50,maxY = 100)
 	public static final BlockOre oreTopaz = new BlockOre(Items.topaz,1,1,1,4,6).setRareOre();
+	@Reg(Keys.oreTanzanite)
+	@GenOreWorld(times=2,timeRate = 0.8f,size=4,minY = 50,maxY = 100)
+	public static final BlockOre oreTanzanite = new BlockOre(Items.tanzanite,1,1,1,4,6).setRareOre();
 
 	@Reg(Keys.orePolarium)
 //	@GenMeteoWorld
@@ -181,6 +184,8 @@ public class Blocks
 	public static final Block blockOpal = new BlockCompressed(Material.ROCK); // 欧珀块
 	@Reg(Keys.blockTopaz)
 	public static final Block blockTopaz = new BlockCompressed(Material.ROCK); // 托帕石块
+	@Reg(Keys.blockTanzanite)
+	public static final Block blockTanzanite = new BlockCompressed(Material.ROCK); // 坦桑石块
 
 	@Reg(Keys.blockStellariumObsidian)
 	public static final Block blockStellariumObsidian = new Block(Material.ROCK); // 恒星黑曜石