+++ src/main/java/firok/tiths/common/Items.java
@@ -73,6 +73,8 @@ public class Items
 	public static final Item opal = new Item(); // 欧珀
 	@Reg(nameTopaz)
 	public static final Item topaz = new Item(); // 托帕石
+	@Reg(nameTanzanite)
+	public static final Item tanzanite = new Item(); // 坦桑石
 
 	// 金属粒
 	@Reg(Keys.nuggetStellarium)